+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+ * Enhance `Parser` to support a bit better `&&` tokens and C++11 rvalue references ([issue bytedeco/javacpp-presets#160](https://github.com/bytedeco/javacpp-presets/issues/160))
  * Add properties for the `linux-armhf` platform, and pick up `macosx` when `os.name` is `darwin`
  * Fix `NullPointerException` in `Parser` on unexpected forms of function pointers ([issue #70](https://github.com/bytedeco/javacpp/issues/70))
  * Make sure `Generator` produces calls to `sizeof()` and `offsetof()` for all `Pointer` classes with allocators